+++ TwitterJSONTests/TwitterJSONTests.swift
@@ -8,23 +8,54 @@
 
 import UIKit
 import XCTest
-import Alamofire
 import SwiftyJSON
+
 
 class TwitterJSONTests: XCTestCase {
     
+    var apiKey: String?
+    var apiSecret:String?
+    var tj: TwitterJSON?
+    
+    override func setUp() {
+        apiKey = "P7JKEueHsWkf8lpqxEN4f1amw"
+        apiSecret = "4onJJkjrSEaySwT6pN1reT9RtpnzGSGeZfBszYm4GP0vVbznbD"
+        tj = TwitterJSON(apiKey: apiKey!, apiSecret: apiSecret!)
+    }
+    
     func testExample() {
-        var tweetJSON = TwitterJSON(apiKey: "xxx", apiSecret: "xxx")
-//        XCTAssertEqual(tweetJSON.apiKey, "xxx")
-//        XCTAssertEqual(tweetJSON.apiSecret, "xxx")
+        XCTAssertEqual(tj!.apiKey, apiKey!)
+        XCTAssertEqual(tj!.apiSecret, apiSecret!)
     }
     
-    func testMakeBearerTokenCredentials() {
-        var tweetJSON = TwitterJSON(apiKey: "xxx", apiSecret: "xxx")
-        tweetJSON.getBearerToken {
-            (bearerToken) -> Void in
-            XCTAssertNotNil(bearerToken)
-        }
+    func testGetBearerToken() {
+        let expectation = expectationWithDescription("Get Bearer Token")
+        let expectedBearerToken = "AAAAAAAAAAAAAAAAAAAAALSQgwAAAAAAfYtOdeJI%2BWbLZyi8fcyfy4OieF4%3DhiTgt5j3plYuaZX7rBJ1zMJzAyWmD0bxFSCjVNa7hKR4hXO0w6"
+
+        tj!.getBearerToken({ (bearerToken) -> Void in
+            XCTAssertEqual(bearerToken, expectedBearerToken)
+            expectation.fulfill()
+        })
+        
+        waitForExpectationsWithTimeout(5, handler: { error in
+            XCTAssertNil(error, "Error")
+        })
     }
     
+    func testPerformDataRequestForURL() {
+        let expectation = expectationWithDescription("Got Tweets")
+        let bearerToken = "AAAAAAAAAAAAAAAAAAAAALSQgwAAAAAAfYtOdeJI%2BWbLZyi8fcyfy4OieF4%3DhiTgt5j3plYuaZX7rBJ1zMJzAyWmD0bxFSCjVNa7hKR4hXO0w6"
+        let apiURL = "https://api.twitter.com/1.1/statuses/user_timeline.json?screen_name=kylegoslan"
+        tj!.performDataRequestForURL(apiURL, bearerToken: bearerToken, completion: { data in
+            for item in data {
+                let id = item.1["id"].int
+                XCTAssertNotNil(id)
+            }
+            expectation.fulfill()
+        })
+        waitForExpectationsWithTimeout(5, handler: { error in
+            XCTAssertNil(error, "Error")
+        })
+    }
+
 }